Move tests that do not modify database
