+++ .circleci/config.yml
@@ -66,15 +66,7 @@ jobs:
       - setup_integration_test
       - run:
           name: Test
-          command: bash -i -c 'npx cypress run --record --config numTestsKeptInMemory=1 --spec cypress/integration/group3/**/*'
-  integration_test_4:
-    executor: integration_test_exec
-    working_directory: ~/repo
-    steps:
-      - setup_integration_test
-      - run:
-          name: Test
-          command: bash -i -c 'npx cypress run --record --config numTestsKeptInMemory=1 --spec cypress/integration/group4/**/*'
+          command: bash -i -c 'npx cypress run --record --config numTestsKeptInMemory=1 --spec cypress/integration/testThatDoNotModifyDatabase/**/*'
   upload_to_s3:
     working_directory: ~/repo
     docker:
@@ -161,19 +153,12 @@ workflows:
               only: /.*/
           requires:
             - lint_unit_test_coverage
-      - integration_test_4:
-          filters:
-            tags:
-              only: /.*/
-          requires:
-            - lint_unit_test_coverage
       # Upload builds for tags to s3. We can add branches later
       - upload_to_s3:
           requires:
             - integration_test_1
             - integration_test_2
             - integration_test_3
-            - integration_test_4
           filters:
             tags:
               only: /.*/
