+++ cypress/integration/group2/starDisabled.ts
