+++ cypress/integration/group2/starEntry.ts
