+++ cypress/integration/group2/starErrorMessage.ts
