+++ cypress/integration/testThatDoNotModifyDatabase/app-spec.ts
@@ -13,10 +13,9 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { resetDB, setTokenUserViewPort } from '../../support/commands';
+import { setTokenUserViewPort } from '../../support/commands';
 
 describe('Dockstore Home', () => {
-  resetDB();
   setTokenUserViewPort();
   beforeEach(() => {
     cy.visit('');
