+++ cypress/integration/testThatDoNotModifyDatabase/curator.ts
@@ -13,15 +13,14 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { resetDB, setTokenUserViewPort } from '../../support/commands';
-describe('Admin UI', () => {
-  resetDB();
+import { setTokenUserViewPort } from '../../support/commands';
+describe('Curator UI', () => {
   setTokenUserViewPort();
   beforeEach(() => {
     cy.server();
-    const userObject = { id: 1, username: 'user_A', curator: false, isAdmin: true, name: 'user_A', setupComplete: false};
+    const userObject = { id: 1, username: 'user_A', curator: true, isAdmin: false, name: 'user_A', setupComplete: false};
     cy.route({
-      url: '/users/user',
+      url: '*/users/user',
       method: 'GET',
       status: 200,
       response: userObject
@@ -32,16 +31,16 @@ describe('Admin UI', () => {
     cy.get('[data-cy=dropdown-main]:visible').click();
   });
 
-  describe('Dropdown', () => {
-    it('Admin user sees lock icon', () => {
+  describe('Dropdown for curators', () => {
+    it('Curator user sees edit icon', () => {
       cy.
       get('#dropdown-accounts > .mat-icon')
-        .contains('lock');
+        .contains('edit');
     });
   });
 
   describe('Profile', () => {
-    it('Admin status indicated on profile page', () => {
+    it('Curator status indicated on profile page', () => {
       cy
         .get('#dropdown-accounts')
         .click();
@@ -49,11 +48,11 @@ describe('Admin UI', () => {
         .get('#mat-tab-label-0-1')
         .click();
       cy
-        .get('#account-is-admin')
+        .get('#account-is-curator')
         .should('exist');
       cy
-        .get('#account-is-admin > .mat-icon')
-        .contains('lock');
+        .get('#account-is-curator > .mat-icon')
+        .contains('edit');
     });
   });
 });
