+++ cypress/integration/testThatDoNotModifyDatabase/verification.ts
@@ -1,8 +1,7 @@
 import { Dockstore } from '../../../src/app/shared/dockstore.model';
-import { resetDB, setTokenUserViewPort } from '../../support/commands';
+import { setTokenUserViewPort } from '../../support/commands';
 
 describe('See verification information and logs', () => {
-    resetDB();
     setTokenUserViewPort();
     it('should see logs', () => {
         cy.exec(`curl -X POST "${Dockstore.API_URI}/api/ga4gh/v2/extended/quay.io%2Fgaryluu%2Fdockstore-cgpmap%2Fcgpmap-cramOut/versions/3.0.0-rc8/CWL/tests/..%2Fexamples%2Fcgpmap%2FcramOut%2Ffastq_gz_input.json?platform=Dockstore%20CLI&platform_version=1.6.0&verified=true&metadata=Potato" -H "accept: application/json" -H "Authorization: Bearer imamafakedockstoretoken2"`);
