+++ cypress/integration/testThatDoNotModifyDatabase/workflowsSearch.ts
@@ -13,10 +13,9 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { resetDB, setTokenUserViewPort } from '../../support/commands';
+import { setTokenUserViewPort } from '../../support/commands';
 
 describe('Dockstore workflow list page', () => {
-  resetDB();
   setTokenUserViewPort();
   describe('Select a workflow', () => {
     it('Should be able to go to the workflows search page', () => {
